pyproject.toml
[project]
name = "ginja-ai"
version = "0.1.0"
description = "Ginja AI — Production FastAPI application"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.131.0",
    "pydantic-settings>=2.7.1",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30.0",
    "alembic>=1.15.0",
    "passlib>=1.7.4",
    "bcrypt==4.0.1",
    "python-jose[cryptography]>=3.5.0",
    "authlib>=1.6.8",
]

[project.scripts]
ginja-ai = "main:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "pytest-cov>=7.0.0",
    "pre-commit>=4.5.1",
    "ruff>=0.15.2",
]
